copy deps folder
dealing with package local dependencies can be a hassle.
The solution here is to force package local dependencies
to be in the deps folder. This likely will require some
massaging of the package-lock.json; however it's an ok
tradeoff to me between auto-detection complexity and
offering an escape hatch for local dependencies.

: